# packages/schemas/openapi.yaml
openapi: 3.1.0
info:
  title: Hoopoe Daemon API
  version: 0.1.0
  description: |
    Hoopoe daemon ↔ desktop contract per `plan.md §2.6`.

    Conventions (declared once here, applied across endpoints):

    - **Errors** use RFC 7807 `application/problem+json` (`#/components/schemas/Problem`).
    - **Idempotency** for retryable writes: clients pass `Idempotency-Key` (a stable
      ULID/UUID); the daemon dedupes within `IdempotencyKey.windowSeconds` and replays
      the original response on retry. Required on every non-GET that mutates state.
    - **Schema versions** every persisted entity declares `schemaVersion`. The daemon
      runs migrations at startup with backup + rollback (`plan.md §10.3`).
    - **Capability gating** every UI feature is gated on capability IDs reported by
      `/v1/capabilities` (§2.8). A capability that parses but cannot satisfy a
      contract test is `degraded`, not `ok`.
    - **WS events** the WebSocket envelope is documented as JSON Schema in
      `events/ws-envelope.schema.json` (sidecar; OpenAPI 3.1 cannot model the
      event-stream shape adequately).
    - **Source of truth** this spec is authoritative; both the TS client
      (`packages/schemas/ts/`) and the Go types (`packages/schemas/go/`) are
      generated from it. Hand-maintained duplicate shapes in `apps/desktop` or
      `apps/daemon` are forbidden (`plan.md §3 'Shared'`).
  license:
    name: UNLICENSED
servers:
  - url: http://127.0.0.1:0
...
  - url: http://127.0.0.1:8765
...
security:
  - bearerAuth: []
...
paths:
...
  /v1/jobs:
    get:
      tags: [Jobs]
      operationId: listJobs
      summary: List jobs (queued / running / waiting_approval / terminal).
      parameters:
        - name: status
          in: query
          required: false
          description: Filter by job status (multi-value).
          explode: true
          schema:
            type: array
            items: { $ref: '#/components/schemas/JobStatus' }
        - name: projectId
          in: query
          required: false
          schema: { type: string }
        - name: cursor
          in: query
          required: false
          schema: { $ref: '#/components/schemas/Cursor' }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        '200':
          description: Job list (paginated).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobListResponse' }
        default: { $ref: '#/components/responses/Problem' }
...
  /v1/events/replay:
    get:
      tags: [Events]
      operationId: replayEvents
      summary: Replay events from a sequence cursor for one channel.
      description: |
        Used by the desktop on reconnect (after laptop sleep, daemon
        restart, tunnel re-establishment, or a `_gap` notice on the WS).
        Returns events strictly after `sinceSequence` up to the daemon's
        retention horizon. Past the horizon, the desktop falls back to
        the persisted-log offset reported in the `_gap` notice.
      parameters:
        - name: channel
          in: query
          required: true
          description: Channel selector (e.g., `project:proj_01`).
          schema: { type: string, minLength: 1 }
        - name: sinceSequence
          in: query
          required: true
          description: Last-known sequence number; replay returns sequence > this.
          schema: { type: integer, minimum: 0 }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 1000, default: 200 }
      responses:
        '200':
          description: Replayed event window.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EventReplayResponse' }
        default: { $ref: '#/components/responses/Problem' }

  # ─────────────────────────────────────────────────────────────────────────
  # Tending (§8) — submit an ActionPlan; daemon validates + dry-runs +
  # enforces policy/approvals + executes + verifies postconditions.
  # ─────────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: HMAC-signed session token (`plan.md §5.2`)
      description: |
        30-day bearer minted by exchanging a single-use pairing token at
        `/v1/auth/bootstrap/bearer`. Stored on the desktop via Electron
        `safeStorage.encryptString`. WebSocket connects use a 5-min `wsToken`
        from `/v1/auth/ws-token` instead.
...
  responses:
    Problem:
      description: RFC 7807 problem+json error envelope.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Problem:
      type: object
      description: |
        RFC 7807 problem+json. `type` is a daemon-stable URN; `code` is a short
        machine-readable identifier (e.g., `auth.token_expired`,
        `capability.missing`). `instance` is a per-incident URI for support
        deep-links into Diagnostics.
      required: [type, title, status, code]
      properties:
        type:
          type: string
          format: uri-reference
          description: Stable problem-type URN (e.g., `urn:hoopoe:auth/token-expired`).
        title:
          type: string
          description: Short, human-readable summary.
        status:
          type: integer
          minimum: 100
          maximum: 599
          description: Mirrors HTTP status.
        code:
          type: string
          description: Machine-readable error code (dot-separated namespace).
          pattern: '^[a-z][a-z0-9_]*(\.[a-z][a-z0-9_]*)*$'
        detail:
          type: string
          description: Human-readable explanation; redacted for sensitive fields.
        instance:
          type: string
          format: uri-reference
          description: Per-incident URI; `urn:hoopoe:incident:<id>` is canonical.
        requestId:
          type: string
          description: Echoes `X-Request-Id` if provided.
        correlationId:
          type: string
          description: Opaque correlation handle (matches audit-log entries).
        retryable:
          type: boolean
          description: True if the client may safely retry with the same Idempotency-Key.
        retryAfterSeconds:
          type: integer
          minimum: 0
          description: Optional hint for `Retry-After`.
        capability:
          type: string
          description: |
            For `capability.missing` / `capability.degraded` problems, the
            capability ID that the operation required.
        approvalId:
          type: string
          description: For `approval.required` problems, the queued approval ID.
    SchemaVersion:
      type: integer
      minimum: 1
      description: Monotonically increasing schema version for a persisted shape.
    Cursor:
      type: string
      description: |
        Opaque pagination cursor. Use the `next` cursor from a list response
        verbatim; do not parse it. Cursors are stable for ≥24h.
    PageMeta:
      type: object
      required: [hasMore]
      properties:
        hasMore: { type: boolean }
        next: { $ref: '#/components/schemas/Cursor' }
        total:
          type: integer
          description: Optional total count; expensive list endpoints omit this.
    Actor:
      type: object
      description: Who initiated an action — used in audit + event envelopes.
      required: [kind]
      properties:
        kind:
          type: string
          enum: [user, agent, scheduler, system, dcg, approval]
        id:
          type: string
          description: Stable identifier within the actor's namespace.
        displayName:
          type: string
          description: Optional UI-only label; never used for authorization.

    # ─────────────────────────────────────────────────────────────────────────
    # System endpoints
    # ─────────────────────────────────────────────────────────────────────────
...
    JobStatus:
      type: string
      enum: [queued, running, waiting_approval, canceling, succeeded, failed, interrupted]
    ArtifactRef:
      type: object
      description: |
        Stable handle to a daemon-managed artifact (logs, plan outputs,
        conversion traces, health snapshots, review findings, bootstrap
        logs). Resolved via the artifact registry on demand.
      required: [id, kind]
      properties:
        id: { type: string }
        kind:
          type: string
          enum:
            - log
            - plan_artifact
            - conversion_trace
            - health_snapshot
            - review_finding
            - bootstrap_log
            - misc
        sizeBytes: { type: integer, minimum: 0 }
        contentType: { type: string }
        sha256: { type: string }
        createdAt: { type: string, format: date-time }
    Job:
      type: object
      description: |
        One long-running job in the daemon's registry (§2.7). `argvDigest`
        and `envDigest` are blake3 hashes — full argv/env are recorded in the
        artifact registry, redacted on the way out.
      required: [id, type, status, schemaVersion]
      properties:
        schemaVersion: { $ref: '#/components/schemas/SchemaVersion' }
        id: { type: string, description: 'Sortable ULID for the job.' }
        type:
          type: string
          description: |
            Job kind (e.g., `bootstrap.acfs`, `plan.candidate.generate`,
            `swarm.spawn`, `health.snapshot`, `tend.swarm`,
            `build.run`, `test.run`).
        status: { $ref: '#/components/schemas/JobStatus' }
        projectId: { type: string }
        actor: { $ref: '#/components/schemas/Actor' }
        argvDigest: { type: string, description: 'Hash of the normalized argv.' }
        envDigest: { type: string, description: 'Hash of the relevant env keys.' }
        cwd: { type: string }
        startedAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time }
        durationMs: { type: integer, minimum: 0 }
        exitCode: { type: integer }
        failureFingerprint:
          type: string
          description: |
            Stable hash of the failure signature (test name, stack frames),
            used to detect repeated failures across reruns (§2.7).
        artifacts:
          type: array
          items: { $ref: '#/components/schemas/ArtifactRef' }
        approvalId:
          type: string
          description: 'Set when status is `waiting_approval`.'
    JobListResponse:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Job' }
        page: { $ref: '#/components/schemas/PageMeta' }
...
    WsEventEnvelope:
      type: object
      description: |
        Daemon → client envelope for every project / swarm / activity event.
        `sequence` is monotonic per `channel`; clients persist the last seen
        and pass it back on reconnect for replay.
      required: [eventId, schemaVersion, channel, type, sequence, time]
      properties:
        eventId: { type: string, description: 'ULID for the event.' }
        schemaVersion: { $ref: '#/components/schemas/SchemaVersion' }
        channel:
          type: string
          description: 'e.g., `project:proj_01`, `swarm:sw_01`.'
        type:
          type: string
          description: |
            Channel-scoped event type (e.g., `bead.changed`, `agent.idle`,
            `job.status.changed`). The daemon publishes the per-channel
            registry; clients tolerate unknown types (forward-compat).
        sequence:
          type: integer
          minimum: 0
          description: 'Monotonic per channel.'
        time: { type: string, format: date-time }
        actor: { $ref: '#/components/schemas/Actor' }
        causationId:
          type: string
          description: 'ID of the command/action that caused this event.'
        correlationId:
          type: string
          description: 'Cross-event handle (e.g., swarm session ID).'
        data:
          description: |
            Event-type-specific payload. Schema per `type`; consumers should
            switch on `type` and validate against the published per-type schema.
            Unknown types should not fail-closed (forward-compat).
...
    EventReplayResponse:
      type: object
      required: [channel, events, latestSequence]
      properties:
        channel: { type: string }
        events:
          type: array
          items: { $ref: '#/components/schemas/WsEventEnvelope' }
        latestSequence:
          type: integer
          minimum: 0
          description: 'Highest sequence currently retained for this channel.'
        retentionHorizon:
          type: integer
          minimum: 0
          description: |
            Lowest sequence still available via replay. Sequences below this
            are reachable only via the persisted-log offset on disk.
        truncated:
          type: boolean
          description: 'True iff `limit` capped the result (more events available).'

    # ─────────────────────────────────────────────────────────────────────────
    # ActionPlan (§8.3.1) — typed agent action surface, mirrored to
    # `tending-actions.yaml`. The closed `kind` set is canonical there;
    # CommandSpec.kind is open by design (other callers may add new kinds).
    # ─────────────────────────────────────────────────────────────────────────
...